requirements.txt
py-cord>=2.4.0
python-dotenv>=1.0.0
aiosqlite>=0.19.0
aiohttp>=3.9.0
audioop-lts>=0.2.1    # backport of audioop removed in Python 3.13+
certifi>=2024.0.0     # SSL certificates (required on macOS with python.org Python)
